+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Features
 
+- Make generated destruct branch bodies interactive for clients with snippet edit
+  support. (#2204, @rgrinberg)
 - Offer whole-call completions for OCaml functions with multiple labelled or
   optional arguments, with editable slots for all arguments. (#2199, @rgrinberg)
 - Make generated construct completions interactive for clients with snippet support.